requirements.txt
# Web 框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# AI 和机器学习
openai==1.3.0
transformers==4.35.0
torch>=2.0.0

# 数据处理
pandas==2.1.0
numpy==1.24.0

# 开发工具
python-dotenv==1.0.0
pytest==7.4.0
black==23.0.0